refactor(lab7): fs/ds, fs/fwd
