+++ include/kernel/fs/fs.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "fs/ds.hpp"
+#include "fs/files.hpp"
+#include "fs/vfs.hpp"
