+++ include/kernel/fs/vfs.hpp
@@ -1,140 +1,6 @@
 #pragma once
 
-#include "ds/bitmask_enum.hpp"
-#include "ds/list.hpp"
-#include "util.hpp"
-
-class Vnode;
-class File;
-class FileSystem;
-
-enum filetype {
-  kDir = 1,
-  kFile,
-};
-
-enum fcntl {
-  O_RDONLY = 00000000,
-  O_WRONLY = 00000001,
-  O_RDWR = 00000002,
-  O_ACCMODE = 00000003,
-  O_CREAT = 00000100,
-
-  MARK_AS_BITMASK_ENUM(O_CREAT)
-};
-
-enum seek_type {
-  SEEK_SET = 0,
-};
-
-class Vnode {
-  struct child {
-    char* name;
-    Vnode* node;
-    child(const char* name, Vnode* node);
-    child(const child&);
-    child& operator=(const child&);
-    ~child();
-    bool operator==(const char* name) const;
-  };
-
- private:
-  // TODO: ref cnt
-  Vnode *_parent = nullptr, *_prev = nullptr;
-  list<child> _childs{};
-
- protected:
-  list<child>::iterator find_child(const char* name);
-  int set_child(const char* name, Vnode* vnode);
-  int del_child(const char* name);
-
- public:
-  int link(const char* name, Vnode* vnode);
-
-  const filetype type;
-
-  void set_parent(Vnode* parent);
-  Vnode(filetype type);
-
-  Vnode* parent() const {
-    return _parent;
-  }
-  const list<child>& childs() const {
-    return _childs;
-  }
-
-  bool isDir() const {
-    return type == kDir;
-  }
-  bool isFile() const {
-    return type == kFile;
-  }
-
-  virtual ~Vnode() = default;
-  int lookup(const char* component_name, Vnode*& vnode);
-  const char* lookup(const char* component_name);
-  int mount(const char* component_name, Vnode* new_vnode);
-  virtual long size() const;
-  virtual int create(const char* component_name, Vnode*& vnode);
-  virtual int mkdir(const char* component_name, Vnode*& vnode);
-  virtual int open(const char* component_name, File*& file, fcntl flags);
-  virtual int close(File* file);
-  template <typename F>
-  int _open(const char* component_name, File*& file, fcntl flags) {
-    file = new F{component_name, this, flags};
-    if (file == nullptr)
-      return -1;
-    return 0;
-  }
-};
-
-// file handle
-class File {
- protected:
-  const char* _name;
-  Vnode* vnode;
-  size_t f_pos = 0;  // RW position of this file handle
-  const fcntl flags;
-
- public:
-  const bool can_seek = true;
-
-  const char* name() const {
-    return _name;
-  }
-  size_t size() const {
-    return vnode->size();
-  }
-  fcntl accessmode() const {
-    return flags & O_ACCMODE;
-  }
-  bool canRead() const {
-    return accessmode() == O_RDONLY or accessmode() == O_RDWR;
-  }
-  bool canWrite() const {
-    return accessmode() == O_WRONLY or accessmode() == O_RDWR;
-  }
-
-  File(const char* name, Vnode* vnode, fcntl flags)
-      : _name(name), vnode{vnode}, flags{flags} {}
-  virtual ~File() = default;
-
-  virtual int write(const void* buf, size_t len);
-  int _read(const void* data, void* buf, size_t len);
-  virtual int read(void* buf, size_t len);
-  virtual long lseek64(long offset, seek_type whence);
-  int close();
-};
-
-class FileSystem {
- public:
-  const char* const name;
-  FileSystem* next = nullptr;
-
-  FileSystem(const char* name) : name(name) {}
-
-  virtual Vnode* mount();
-};
+#include "fs/fwd.hpp"
 
 extern Vnode* root_node;
 void init_vfs();
