+++ src/kernel/shell/cmd_ls.cpp
@@ -1,4 +1,4 @@
-#include "fs/files.hpp"
+#include "fs/fs.hpp"
 #include "io.hpp"
 #include "shell/cmd.hpp"
